+++ tests/fixtures/regression/kv/seed__kv.cloesce.ts
@@ -1,5 +1,17 @@
 import {
-    KV, KValue, WranglerEnv, DataSource, IncludeTree, KeyParam, Model, PrimaryKey, GET
+    KV,
+    KValue,
+    WranglerEnv,
+    DataSource,
+    IncludeTree,
+    KeyParam,
+    Model,
+    PrimaryKey,
+    Inject,
+    DeepPartial,
+    Orm,
+    POST,
+    CRUD
 } from "cloesce/backend";
 import { D1Database } from "@cloudflare/workers-types";
 
@@ -13,7 +25,8 @@ export class Env {
     otherNamespace: KVNamespace;
 }
 
-@CRUD(["SAVE", "GET"])
+// TODO: CRUD FOR KV
+@CRUD
 @Model
 export class PureKVModel {
     @KeyParam
@@ -30,9 +43,19 @@ export class PureKVModel {
         data: {},
         otherData: {}
     };
+
+    @POST
+    static async post(@Inject env: Env, id: string, data: unknown, otherData: string) {
+        const kvKey = `path/to/data/${id}`;
+        await env.namespace.put(kvKey, JSON.stringify(data));
+
+        const otherKvKey = `path/to/other/${id}`;
+        await env.otherNamespace.put(otherKvKey, otherData);
+    }
 }
 
-@CRUD(["SAVE", "GET", "LIST"])
+// TODO: CRUD FOR KV
+@CRUD()
 @Model
 export class D1BackedModel {
     @PrimaryKey
@@ -47,13 +70,16 @@ export class D1BackedModel {
     @KV("d1Backed/{id}/{keyParam}/{someColumn}/{someOtherColumn}", "namespace")
     kvData: KValue<unknown>;
 
-    @GET
-    instanceMethod(): D1BackedModel {
-        if (this.kvData === undefined) {
-            throw new Error("kvData is undefined");
-        }
+    @POST
+    static async post(@Inject env: Env, model: DeepPartial<D1BackedModel>) {
+        const orm = Orm.fromEnv(env);
+        const id = await orm.upsert(D1BackedModel, model, {});
 
-        return this;
+        const newModel = (await orm.get(D1BackedModel, id.unwrap(), {})).unwrap();
+
+        // upload kvData
+        const kvKey = `d1Backed/${newModel.id}/${newModel.keyParam}/${newModel.someColumn}/${newModel.someOtherColumn}`;
+        await env.namespace.put(kvKey, JSON.stringify(newModel.kvData));
     }
 }
